# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Constellation API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/views/passes": {
      "get": {
        "tags": [
          "views"
        ],
        "summary": "Recent clear passes over a place or spot",
        "description": "Clear Sentinel-2 passes over the place (or the square around the pin) in the period.",
        "operationId": "view_passes_api_views_passes_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "4m",
                "1y",
                "2y",
                "5y"
              ],
              "type": "string",
              "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month).",
              "default": "4m",
              "title": "Period"
            },
            "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month)."
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 85,
                  "minimum": -85
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          },
          {
            "name": "place_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "A saved place: its outline.",
              "title": "Place Id"
            },
            "description": "A saved place: its outline."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$.",
              "title": "X-User-Id"
            },
            "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ViewPass"
                  },
                  "title": "Response View Passes Api Views Passes Get"
                }
              }
            }
          },
          "400": {
            "description": "Give either `place_id`, or both `lat` and `lon`."
          },
          "404": {
            "description": "Unknown place, or no imagery here (offline data or no clear pass)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/views": {
      "get": {
        "tags": [
          "views"
        ],
        "summary": "Render one band of a place or spot",
        "description": "One band (photo, greenness, water, bare) from the given pass or the latest clear one.\nA place is drawn inside its own outline (transparent outside); a pin gets a 2 km square.\nRendered once, then cached. No agent, no cost.",
        "operationId": "view_image_api_views_get",
        "parameters": [
          {
            "name": "band",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "photo",
                "greenness",
                "water",
                "bare"
              ],
              "type": "string",
              "title": "Band"
            }
          },
          {
            "name": "scene",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scene"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "4m",
                "1y",
                "2y",
                "5y"
              ],
              "type": "string",
              "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month).",
              "default": "4m",
              "title": "Period"
            },
            "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month)."
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 85,
                  "minimum": -85
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          },
          {
            "name": "place_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "A saved place: its outline.",
              "title": "Place Id"
            },
            "description": "A saved place: its outline."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$.",
              "title": "X-User-Id"
            },
            "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewImage"
                }
              }
            }
          },
          "400": {
            "description": "Give either `place_id`, or both `lat` and `lon`."
          },
          "404": {
            "description": "Unknown place, or no imagery here (offline data or no clear pass)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/views/prefetch": {
      "post": {
        "tags": [
          "views"
        ],
        "summary": "Render every band of every pass ahead of time",
        "description": "Queue rendering of photo, greenness, water and bare ground for every clear pass in the\nperiod, so flicking through dates is instant. Runs in the background; a place's recent\nwindow is already queued when it is saved.",
        "operationId": "view_prefetch_api_views_prefetch_post",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "4m",
                "1y",
                "2y",
                "5y"
              ],
              "type": "string",
              "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month).",
              "default": "4m",
              "title": "Period"
            },
            "description": "How far back: `4m` (the latest clear passes, default), or `1y`, `2y`, `5y` (the clearest pass of each month)."
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 85,
                  "minimum": -85
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          },
          {
            "name": "place_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "A saved place: its outline.",
              "title": "Place Id"
            },
            "description": "A saved place: its outline."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$.",
              "title": "X-User-Id"
            },
            "description": "Optional demo user id (default \"demo\"), ^[a-z0-9][a-z0-9_-]{0,31}$."
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefetchResult"
                }
              }
            }
          },
          "400": {
            "description": "Give either `place_id`, or both `lat` and `lon`."
          },
          "404": {
            "description": "Unknown place, or no imagery here (offline data or no clear pass)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrefetchResult": {
        "properties": {
          "passes": {
            "type": "integer",
            "title": "Passes",
            "description": "Clear passes in the period."
          },
          "images": {
            "type": "integer",
            "title": "Images",
            "description": "Images that will be cached: passes times bands."
          },
          "queued": {
            "type": "boolean",
            "title": "Queued",
            "description": "False when the same prefetch is already running."
          }
        },
        "type": "object",
        "required": [
          "passes",
          "images",
          "queued"
        ],
        "title": "PrefetchResult",
        "description": "What a prefetch will render in the background."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ViewImage": {
        "properties": {
          "band": {
            "type": "string",
            "enum": [
              "photo",
              "greenness",
              "water",
              "bare"
            ],
            "title": "Band"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Served by GET /api/layers/...; relative to the API origin."
          },
          "bounds": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bounds",
            "description": "[west, south, east, north]"
          },
          "scene": {
            "type": "string",
            "title": "Scene"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "satellite": {
            "type": "string",
            "title": "Satellite"
          },
          "cloud": {
            "type": "integer",
            "title": "Cloud"
          }
        },
        "type": "object",
        "required": [
          "band",
          "url",
          "bounds",
          "scene",
          "date",
          "satellite",
          "cloud"
        ],
        "title": "ViewImage",
        "description": "One band of one pass, rendered as a PNG pinned to WGS84 bounds."
      },
      "ViewPass": {
        "properties": {
          "scene": {
            "type": "string",
            "title": "Scene"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "satellite": {
            "type": "string",
            "title": "Satellite"
          },
          "cloud": {
            "type": "integer",
            "title": "Cloud",
            "description": "Cloud over the square, percent."
          }
        },
        "type": "object",
        "required": [
          "scene",
          "date",
          "satellite",
          "cloud"
        ],
        "title": "ViewPass",
        "description": "A recent clear Sentinel-2 pass over the spot."
      },
...
    }
  }
}